spectrogram refactor
